# perf: cache jest fn call parsing results