@SimenB 2022-08-07
> (@SimenB we should make sure that `pref` actually triggers a release before merging this 😅)

It should trigger a patch release, looking at https://github.com/semantic-release/commit-analyzer/blob/ad3079b515be8266da02cc6dc64f765dd8aaf509/lib/default-release-rules.js#L12